Fix Jool.
Jool can not be modified in current versions of Kopernicus. It has been replaced with a clone "UL_Jool" with the exact same displayName and properties as the original UL_Jool. UBI support from my fork allows Jool's moons and asteroids to be totally supported, however visual mods will need patches to specify the new Jool.